new files

+++ src/morrf_awa/morrf_awa/subtree.h
@@ -41,7 +41,7 @@ public:
 class RRTree {
 public:
     enum TREE_TYPE{ UNKNOWN, SUBPROBLEM, REFERENCE };
-    RRTree( MORRF* parent, unsigned int objective_num, std::vector<double> weight, unsigned int index);
+    RRTree( MORRF* parent, unsigned int objective_num, std::vector<float> weight, unsigned int index);
 
     RRTNode* init( POS2D start, POS2D goal );
     RRTNode* create_new_node( POS2D pos );
@@ -84,7 +84,7 @@ public:
 
 class ReferenceTree : public RRTree {
 public:
-    ReferenceTree( MORRF* parent, unsigned int objective_num, std::vector<double> weight, unsigned int index );
+    ReferenceTree( MORRF* parent, unsigned int objective_num, std::vector<float> weight, unsigned int index );
 
     virtual void attach_new_node( RRTNode* p_node_new, RRTNode* p_nearest_node, std::list<RRTNode*> near_nodes );
     virtual void rewire_near_nodes( RRTNode* p_node_new, std::list<RRTNode*> near_nodes );
@@ -98,7 +98,7 @@ protected:
 
 class SubproblemTree : public RRTree {
 public:
-    SubproblemTree( MORRF* parent, unsigned int objective_num, std::vector<double> weight, unsigned int index );
+    SubproblemTree( MORRF* parent, unsigned int objective_num, std::vector<float> weight, unsigned int index );
 
     virtual void attach_new_node( RRTNode* p_node_new, RRTNode* p_nearest_node, std::list<RRTNode*> near_nodes );
     virtual void rewire_near_nodes( RRTNode* p_node_new, std::list<RRTNode*> near_nodes );
